+++ core/models/simulation.py
@@ -381,6 +381,56 @@ class HXSimulationResult:
     def outside_dp_local(self) -> float:
         return self.final_result.outside_dp_local
 
+    # -- Tube-sheet entrance/exit pressure drop (v0.5.6) --------------------
+
+    @property
+    def inside_dp_straight_tube_friction(self) -> float:
+        return self.final_result.inside_dp_straight_tube_friction
+
+    @property
+    def inside_dp_straight_tube_acceleration(self) -> float:
+        return self.final_result.inside_dp_straight_tube_acceleration
+
+    @property
+    def inside_dp_straight_tubes(self) -> float:
+        return self.final_result.inside_dp_straight_tubes
+
+    @property
+    def inside_dp_tube_entrances(self) -> float:
+        return self.final_result.inside_dp_tube_entrances
+
+    @property
+    def inside_dp_tube_exits(self) -> float:
+        return self.final_result.inside_dp_tube_exits
+
+    @property
+    def tube_path_type(self):
+        return self.final_result.tube_path_type
+
+    @property
+    def entrance_count(self) -> int:
+        return self.final_result.entrance_count
+
+    @property
+    def exit_count(self) -> int:
+        return self.final_result.exit_count
+
+    @property
+    def pass_boundary_method(self) -> str:
+        return self.final_result.pass_boundary_method
+
+    @property
+    def pass_boundary_states(self):
+        return self.final_result.pass_boundary_states
+
+    @property
+    def entrance_results(self):
+        return self.final_result.entrance_results
+
+    @property
+    def exit_results(self):
+        return self.final_result.exit_results
+
 
 # ---------------------------------------------------------------------------
 # Driver
@@ -521,6 +571,8 @@ def run_simulation(
             flow_area_per_pass=hx.bundle.internal_flow_area_per_pass,
             hydraulic_diameter=hx.bundle.internal_hydraulic_diameter,
             hydraulic_length_total=hx.bundle.internal_length_total,
+            n_tube_passes=hx.bundle.n_passes_tube,
+            tube_path_type=hx.bundle.tube_path_type,
             provider=inside.provider,
             temperature_in=inside.T_in,
             temperature_out=T_out_inside_calc,